+++ IMGFR.txt
@@ -1,53 +1,1 @@
-Airplane.jpg
-ApplePicking.jpg
-ArtMuseum.jpg
-AvocadoToast.jpg
-BBQ.jpg
-BabyAnnouncement.jpg
-Basketball.jpg
-BathroomStall.jpg
-BirthdayParty.jpg
-Bonfire.jpg
-BubbleBath.jpg
-Cat.jpg
-Chimpanzee.jpg
-Chinese.jpg
-CompactCar.jpg
-Dog.jpg
-Dolphin.jpg
-EngagementRing.jpg
-FeetOnBeach.jpg
-Fireworks.jpg
-Funeral.jpg
-Giraffe.jpg
-Graduation.jpg
-Gym.jpg
-Gyro.jpg
-Hamburger.jpg
-Hockey.jpg
-Horse.jpg
-HotDogs.jpg
-HouseParty.jpg
-IceCream.jpg
-Kittens.jpg
-London.jpg
-Minivan.jpg
-MovieTheater.jpg
-Nascar.jpg
-NewYorkCity.jpg
-Office.jpg
-Parade.jpg
-Paris.jpg
-PoolParty.jpg
-RomanticGetaway.jpg
-Russia.jpg
-Segway.jpg
-Spider.jpg
-SubwayCar.jpg
-Train.jpg
-Vegas.jpg
-Volleyball.jpg
-Voting.jpg
-Wedding.jpg
-Wrestling.jpg
-WritingDesk.jpg
+test.jpg
