+++ README.md
@@ -5,13 +5,10 @@ It does not mod the tumbnails as of the last time i updated this readme. <br />
 # How do i use this?
 Simple! <br />
 1.Make sure you have python is installed, and its on your [PATH](https://datatofish.com/add-python-to-windows-path/ "Adding Python to PATH"). <br />
-2.Run Setup.bat OR in cmd, run this: pip install Pillow <br />
+2.Run Setup.bat OR in cmd, run this: pip install -r requirements.txt <br />
 3.Put the images you want the program to randomly replace in Replacement Images. <br />
-4.Make sure you backup all of these replacement images in the backup folder, or your own. <br />
-5.Double Click Main.py OR run it in cmd (makes reading the console easier.) <br />
+5.Double Click GUI.py<br />
 
 # Crashed? and other stuff.
-Whenever there are no more replacement images, it just crashes! <br />
-It is completly normal. <br />
-If it crashes when there are still images in that folder, leave a [github issue](https://github.com/weegeeday/Jackbox-sti-script/issues/new/choose). <br />
+If it crashes, leave a [github issue](https://github.com/weegeeday/Jackbox-sti-script/issues/new/choose). <br />
 If you have any requests, suggestions or features youd want, make a [github issue](https://github.com/weegeeday/Jackbox-sti-script/issues/new/choose) about it! <br />
